refactor: move getPlayerInventoryQuantities to its own file
